+++ fronts-client/src/components/card/CardContent.tsx
