+++ fronts-client/src/components/card/CardHeading.tsx
