+++ fronts-client/src/components/card/CardQuote.tsx
