+++ fronts-client/src/components/icons/ColoredQuote.tsx
